+++ tests/AssertTest.php
@@ -61,6 +61,104 @@ final class AssertTest extends TestCase
         $this->assertThrowsException(function () {
             throw new \RuntimeException("abc", 1);
         }, \RuntimeException::class, "abc", 1);
+        $this->assertOutput(function () {
+            echo "abc";
+        }, "abc");
+    }
+
+    /**
+     * Test failures of assertions
+     */
+    public function testAssertionFailures(): void
+    {
+        $this->assertOutput(function () {
+            $this->testResult("abc", false);
+        }, "Test 1 failed. abc\n");
+        $this->assertOutput(function () {
+            $this->assert(false);
+        }, "Test 3 failed. The assertion is not true.\n");
+        $this->assertOutput(function () {
+            $this->assert(false, "abc");
+        }, "Test 5 failed. abc\n");
+        $this->assertOutput(function () {
+            $this->assertSame("abc", "def");
+        }, "Test 7 failed. The value is not abc but def.\n");
+        $this->assertOutput(function () {
+            $this->assertNotSame("abc", "abc");
+        }, "Test 9 failed. The value is abc.\n");
+        $this->assertOutput(function () {
+            $this->assertGreaterThan(2, 1);
+        }, "Test 11 failed. 1 is not greater than 2.\n");
+        $this->assertOutput(function () {
+            $this->assertLessThan(1, 2);
+        }, "Test 13 failed. 2 is not less than 1.\n");
+        $this->assertOutput(function () {
+            $this->assertTrue(false);
+        }, "Test 15 failed. The value is not true.\n");
+        $this->assertOutput(function () {
+            $this->assertTruthy(false);
+        }, "Test 17 failed. The expression is not true.\n");
+        $this->assertOutput(function () {
+            $this->assertFalse(true);
+        }, "Test 19 failed. The value is not false.\n");
+        $this->assertOutput(function () {
+            $this->assertFalsey(true);
+        }, "Test 21 failed. The expression is not false.\n");
+        $this->assertOutput(function () {
+            $this->assertNull("abc");
+        }, "Test 23 failed. The value is not null.\n");
+        $this->assertOutput(function () {
+            $this->assertNotNull(null);
+        }, "Test 25 failed. The value is null.\n");
+        $this->assertOutput(function () {
+            $this->assertContains("abc", "def");
+        }, "Test 27 failed. abc is not in the variable.\n");
+        $this->assertOutput(function () {
+            $this->assertContains("abc", [1, 2, ]);
+        }, "Test 29 failed. abc is not in the variable.\n");
+        $this->assertOutput(function () {
+            $this->assertContains(["abc", ], "abc");
+        }, "Test 31 failed. (array) is not in the variable.\n");
+        $this->assertOutput(function () {
+            $this->assertNotContains("abc", "abc");
+        }, "Test 33 failed. abc is in the variable.\n");
+        $this->assertOutput(function () {
+            $this->assertNotContains("abc", ["abc", "def", ]);
+        }, "Test 35 failed. abc is in the variable.\n");
+        $this->assertOutput(function () {
+            $this->assertNotContains(["abc", "def", ], "abc");
+        }, "Test 37 failed. (array) is not in the variable.\n");
+        $this->assertOutput(function () {
+            $this->assertCount(1, ["abc", "def", ]);
+        }, "Test 39 failed. Count of the variable is 2.\n");
+        $this->assertOutput(function () {
+            $this->assertNotCount(2, ["abc", "def", ]);
+        }, "Test 41 failed. Count of the variable is 2.\n");
+        $this->assertOutput(function () {
+            $this->assertType("string", ["abc", "def", ]);
+        }, "Test 43 failed. The variable is array.\n");
+        $this->assertOutput(function () {
+            $this->assertType(static::class, new stdClass());
+        }, "Test 45 failed. The variable is instance of stdClass.\n");
+        $this->assertOutput(function () {
+            $this->assertThrowsException(function () {
+            }, \Exception::class);
+        }, "Test 47 failed. The code does not throw any exception.\n");
+        $this->assertOutput(function () {
+            $this->assertThrowsException(function () {
+                throw new \RuntimeException();
+            }, \LogicException::class);
+        }, "Test 49 failed. The code does not throw LogicException but RuntimeException.\n");
+        $this->assertOutput(function () {
+            $this->assertThrowsException(function () {
+                throw new \RuntimeException("def");
+            }, \RuntimeException::class, "abc");
+        }, "Test 51 failed. The code does not throw an exception with message 'abc' but 'def'.\n");
+        $this->assertOutput(function () {
+            $this->assertThrowsException(function () {
+                throw new \RuntimeException("abc", 2);
+            }, \RuntimeException::class, "abc", 1);
+        }, "Test 53 failed. The code does not throw an exception with code 1 but 2.\n");
     }
 
     /**